local/requirements.txt
python-dotenv==1.0.0
slack-sdk==3.27.1
psycopg2-binary==2.9.9